+++ compiler-tests/src/test/data/dump/fir/aggregation/FirHintGenerationWorks.kt
@@ -1,0 +1,12 @@
+// GENERATE_JVM_CONTRIBUTION_HINTS_IN_FIR
+
+@ContributesTo(AppScope::class)
+interface ContributedInterface1
+
+@ContributesTo(Unit::class)
+interface ContributedInterface2
+
+// Repeated
+@ContributesTo(AppScope::class)
+@ContributesTo(Unit::class)
+interface ContributedInterface3