+++ compiler-tests/src/test/kotlin/dev/zacsweers/metro/compiler/MetroDirectives.kt
@@ -10,6 +10,10 @@ object MetroDirectives : SimpleDirectivesContainer() {
   val ENABLE_TOP_LEVEL_FUNCTION_INJECTION by directive("Enable top-level function injection.")
   val DISABLE_TRANSFORM_PROVIDERS_TO_PRIVATE by
     directive("Disables automatic transformation of providers to be private.")
+  val GENERATE_JVM_CONTRIBUTION_HINTS_IN_FIR by
+    directive(
+      "Enable/disable generation of contribution hint generation in FIR for JVM compilations types."
+    )
   val PUBLIC_PROVIDER_SEVERITY by
     enumDirective<MetroOptions.DiagnosticSeverity>(
       "Control diagnostic severity reporting of public providers."