+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/fir/Keys.kt
@@ -124,4 +124,8 @@ internal object Keys {
   data object ContributedGraph : GeneratedDeclarationKey() {
     override fun toString() = "ContributedGraph"
   }
+
+  data object ContributionHint : GeneratedDeclarationKey() {
+    override fun toString() = "ContributionHint"
+  }
 }