+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/transformers/ContributionHintIrTransformer.kt
@@ -10,7 +10,8 @@ import dev.zacsweers.metro.compiler.ir.annotationsIn
 import dev.zacsweers.metro.compiler.ir.scopeAnnotations
 import dev.zacsweers.metro.compiler.ir.scopeOrNull
 import dev.zacsweers.metro.compiler.ir.trackFunctionCall
-import dev.zacsweers.metro.compiler.joinSimpleNames
+import dev.zacsweers.metro.compiler.mapNotNullToSet
+import dev.zacsweers.metro.compiler.scopeHintFunctionName
 import org.jetbrains.kotlin.ir.declarations.IrClass
 
 /**
@@ -31,12 +32,12 @@ internal class ContributionHintIrTransformer(
     val contributions =
       declaration.annotationsIn(symbols.classIds.allContributesAnnotations).toList()
 
-    val contributionScopes = contributions.mapNotNullTo(mutableSetOf()) { it.scopeOrNull() }
+    val contributionScopes = contributions.mapNotNullToSet { it.scopeOrNull() }
 
     for (contributionScope in contributionScopes) {
       hintGenerator.generateHint(
         sourceClass = declaration,
-        hintName = contributionScope.joinSimpleNames().shortClassName,
+        hintName = contributionScope.scopeHintFunctionName(),
       )
     }