+++ compiler/src/test/kotlin/dev/zacsweers/metro/compiler/MetroCompilerTest.kt
@@ -101,8 +101,10 @@ abstract class MetroCompilerTest {
                 processor.option(entry.raw.cliOption, generateAssistedFactories)
               MetroOption.ENABLE_TOP_LEVEL_FUNCTION_INJECTION ->
                 processor.option(entry.raw.cliOption, enableTopLevelFunctionInjection)
-              MetroOption.GENERATE_HINT_PROPERTIES ->
-                processor.option(entry.raw.cliOption, generateHintProperties)
+              MetroOption.GENERATE_CONTRIBUTION_HINTS ->
+                processor.option(entry.raw.cliOption, generateContributionHints)
+              MetroOption.GENERATE_JVM_CONTRIBUTION_HINTS_IN_FIR ->
+                processor.option(entry.raw.cliOption, generateJvmContributionHintsInFir)
               MetroOption.TRANSFORM_PROVIDERS_TO_PRIVATE ->
                 processor.option(entry.raw.cliOption, transformProvidersToPrivate)
               MetroOption.SHRINK_UNUSED_BINDINGS ->