+++ gradle-plugin/src/main/kotlin/dev/zacsweers/metro/gradle/MetroPluginExtension.kt
@@ -50,9 +50,22 @@ constructor(layout: ProjectLayout, objects: ObjectFactory, providers: ProviderFa
   public val enableTopLevelFunctionInjection: Property<Boolean> =
     objects.property(Boolean::class.javaObjectType).convention(false)
 
-  /** Enable/disable hint property generation in IR for contributed types. Enabled by default. */
-  public val generateHintProperties: Property<Boolean> =
-    objects.property(Boolean::class.javaObjectType).convention(true)
+  /**
+   * Enable/disable contribution hint generation in IR for contributed types. Enabled by default.
+   *
+   * This does not have a convention default set here as it actually depends on the platform. You
+   * can set a value to force it to one or the other, otherwise if unset it will default to the
+   * default for each compilation's platform type.
+   */
+  public val generateContributionHints: Property<Boolean> =
+    objects.property(Boolean::class.javaObjectType)
+
+  /**
+   * Enable/disable contribution hint generation in FIR for JVM compilations types. Disabled by
+   * default. Requires [generateContributionHints] to be true
+   */
+  public val generateJvmContributionHintsInFir: Property<Boolean> =
+    objects.property(Boolean::class.javaObjectType).convention(false)
 
   /**
    * Enable/disable hint property generation for scoped inject classes. Enabled by default.